test: Fix bug in createAgreement.cy.js
Fixes a bug in the createAgreement.cy.js file where expanding the budget line item and checking the "created by" name would not work properly. This bug was causing the test to fail. The bug has been fixed by adding a click event to the budget line item after checking the "created by" name. This ensures that the test passes successfully.