feat: :sparkles: adds DatePicker (#2197)
* feat: :sparkles: adds DatePicker

feat: Update Prettier configuration and add date picker styling  The Prettier configuration in `.prettierrc.json` has been updated to include consistent indentation and line width settings. Additionally, the date picker styling has been modified to center align the text.  Fixes #2187

2049

* fix: adds refs to open/close accordion

* wip: DatePicker updates for e2e tests

* cleanup

* tests: Update funding summary card values in agreement review

The funding summary card values in the agreement review section have been updated to reflect the correct amounts.

* test: Update createAgreement.cy.js to remove commented out code

The commented out code for selecting the month, day, and year in the createAgreement.cy.js file has been removed. This code is no longer needed as the date is now entered directly using the uswds-date-need-by-date input field.

* test: Fix bug in createAgreement.cy.js

Fixes a bug in the createAgreement.cy.js file where expanding the budget line item and checking the "created by" name would not work properly. This bug was causing the test to fail. The bug has been fixed by adding a click event to the budget line item after checking the "created by" name. This ensures that the test passes successfully.

* fix: Fix bug in agreement review workflow

Fixes a bug in the agreement review workflow where the wrong agreement was being expanded and approved. The bug has been fixed by selecting the last agreement in the table instead of the first one. This ensures that the correct agreement is expanded and the approval process works as expected.

* test: Update funding summary card values in agreement review

* cleanup

* style: Update padding and styling in BudgetLinesForm

The padding and styling in the BudgetLinesForm component have been updated to improve the layout and visual appearance.

* feat: Add validation for needByDate in BudgetLinesForm

The needByDate field in the BudgetLinesForm component now has validation to ensure that the date is in the format MM/DD/YYYY. This prevents users from entering invalid dates and improves data integrity.

* test: Add validation for needByDate in BudgetLinesForm

* refactor: Update needByDate validation in BudgetLinesForm

Update order of validation rules as discussed in UX review 4/18/2024.