feat: ✨ adds validation to DatePicker in EditMode (#2206)
* refactor: Tooltip component

Updayte extension to jsx and update JSDocs comments

* refactor: Update Tooltip component to ensure tooltip availability and cleanup function

* feat: Update BudgetLinesForm to include DatePicker suite

The BudgetLinesForm component has been updated to include the DatePicker suite, allowing for the selection of dates in the MM/DD/YYYY format. This improves the data integrity and user experience.

* feat: Add reset functionality to BudgetLinesForm

The BudgetLinesForm component has been updated to include a reset functionality.

* feat: adds validation for DatePicker in EditMode.

Adds prop isEditMode to BudgetLinesForm and we use that to fire the validation on DatePicker. Also added Vest validation to ensure date is in the future. Lastly, added disabled states to add/edit buttons based on validation status.

* test: Add e2e tests for DatePicker