feat: Add changeRequests Alerts (#2326)
* docs: Update AgreementEditForm.jsx props documentation

Update the props documentation in AgreementEditForm.jsx to make the `selectedAgreementId` prop optional.

* feat: Add change request messages to Alert component

This commit adds the ability to display change request messages in the Alert component. The change request messages are passed as an array to the `changeRequests` prop in the Alert component. This allows for better communication of pending edits to the user.

Changes:
- Modified CreateBLIsAndSCs.hooks.js to include budget line ID in change request messages.
- Updated Alert.jsx to render the change request messages as a list under the main alert message.

Related User Commits:
- feat: Update BLIRow.jsx to render field values in change request messages
- wip: Add changeRequests property to AlertSlice

* docs: Add formData prop to CreateBLIsAndSCs component

* chore: deprecated patchBudgetLineItems.js  in favor of RTK Query

* docs: Update CreateBLIsAndSCs component props documentation

Update the props documentation in CreateBLIsAndSCs.jsx to reflect the correct prop types and descriptions.

* refactor: Rename StatusTag.jsx to StatusTagList.jsx and update exports

This commit renames the file `StatusTag.jsx` to `StatusTagList.jsx` to better reflect its purpose as a list of status tags. The exports have also been updated accordingly.

Changes:
- Renamed `StatusTag.jsx` to `StatusTagList.jsx`.
- Updated exports in `StatusTagList.jsx` to match the new file name.

Related User Commits:
- docs: Update CreateBLIsAndSCs component props documentation

* chore: Update import statement for StatusTagList in AgreementTotalBudgetLinesCard

* revert: Remove changeRequests from Alert component

This refactor removes the `changeRequests` property from the Alert component as it is no longer needed. The code has been simplified and the property has been removed from the state and the JSX rendering.

Changes:
- Removed the `changeRequests` property from the state and JSX rendering in Alert.jsx.

Related User Commits
* docs: Update AgreementDetailsView to use constant for missing value text

* feat: Add AgreementChangesAlert component to AgreementDetails

This commit adds the AgreementChangesAlert component to the AgreementDetails page. The component is conditionally rendered based on the presence of budget line items in review. It also utilizes the useChangeRequests hook to fetch change requests for the agreement. This addition enhances the user experience by providing visibility into any pending change requests for the agreement.

Related User Commits:
- docs: Update AgreementDetailsView to use constant for missing value text

* docs: remove the changeRequests prop

* chore: Remove unused "typeRoots" property from jsconfig.json

* feat: Add isClosable prop to AgreementChangesAlert component

This commit adds the `isClosable` prop to the AgreementChangesAlert component in order to make the alert closable. The prop is optional and defaults to `false`. This enhancement improves the user experience by allowing users to dismiss the alert if desired.

Related User Commits:
- docs: remove the changeRequests prop

* feat: adds AgreementChangesAlert to Agreement details pages

* refactor: Move AgreementChangesAlert component to new directory

The AgreementChangesAlert component has been moved from the "UI/Alert" directory to the "Agreements" directory. This reorganization improves the code structure and makes it easier to locate and manage components related to agreements.

Related User Commits:
- feat: adds AgreementChangesAlert to Agreement details pages
- feat: Add isClosable prop to AgreementChangesAlert component

* refactor: move state for isVisible up to parent

* feat: Add success alert after submitting edits to agreement

This commit adds a success alert after submitting edits to an agreement. The alert notifies the user that their edits have been successfully sent to their Division Director for review. It also provides a redirect URL to the agreement details page for easy navigation.

* refactor: Better defaults for SimpleAlert to keeps things simple

* test: 🧪 adds SimpleAlert component tests

* refactor: useCreateBLIsAndSCs to handle error when updating budget line

* chore: Reset form and set edit mode to false after creating budget lines and subcategories

* test: 🧪 adds tests for AgreementChangeAlert

- adds vi from vitest