feat: Add budget change approval functionality
The code changes in this commit add functionality for approving budget changes. It introduces a new state variable `showModal` and `modalProps` to manage the display of a confirmation modal. When the user clicks on the "Approve" button, the `handleConfirm` function is called, which triggers the `reviewCR` mutation to approve the budget change request. After the approval, a success alert is displayed, and in case of an error, an error alert is shown.