feat: ✨adds Change Requests reviews at card level (#2479)
* chore: fix typo in openapi.yml description

* feat: Add change request review functionality

* test: Add useReviewChangeRequestMutation to change request components

* chore: Fix tabIndex value in ConfirmationModal.jsx

* feat: Add ChangeRequests component for managing change requests

* feat: Add budget change approval functionality

The code changes in this commit add functionality for approving budget changes. It introduces a new state variable `showModal` and `modalProps` to manage the display of a confirmation modal. When the user clicks on the "Approve" button, the `handleConfirm` function is called, which triggers the `reviewCR` mutation to approve the budget change request. After the approval, a success alert is displayed, and in case of an error, an error alert is shown.

* feat: Add handleReviewChangeRequest prop to ChangeRequestList and ReviewCard components

* docs: Make Modal heading and Alert message dynamic per action

* feat: Add budget change approval dynamic functionality

* feat: Update StatusChangeReviewCard and ReviewCard components

This commit adds the `bliToStatus` prop to the `StatusChangeReviewCard` component and the `bliToStatus` parameter to the `ReviewCard` component. This allows for passing the status of the budget line item after the change.

* test: Add forceHover prop to ReviewCard component

This commit adds the `forceHover` prop to the `ReviewCard` component, allowing for the ability to force the hover state. This prop is primarily used for testing purposes.

* feat: Update ReviewCards and variants

This commit adds the `changeMsg` prop to allow for rendering Alerts

* feat: Add 'status' to codesToDisplayText in utils.js

This commit adds the 'status' property to the 'codesToDisplayText' object in the 'utils.js' file. This allows for displaying the status of an item alongside other properties like amount, CAN, and date needed.

* test: Fix agreement list tabs and add budget change test

Updates test for Change Requests and adds a test for Budget Changes

* test: breaks up test

* test: Add BLI Status Change test

The commit adds a new test for BLI Status Change in the `statusChangeRequest.cy.js` file. This test covers the process of creating a test agreement, creating a BLI, submitting it for approval, and deleting the agreement.

* test: Update budget change request amount formatting

This commit updates the amount formatting in the budget change request tests to remove underscores and use plain numbers.

* chore: Update budget change request with new CAN ID

* test: Update budget change request with new CAN ID

We are replacing the api test with UI click thru

* test: request approval test not needed in favor of budgetChangeRequest

* fix: Update CAN ID in statusChangeRequest.cy.js

* fix: Update response status code in backend/openapi.yml

* wip: Add services component creation to budgetChangeRequest.cy.js

* test: updates to UI click thru

* docs: remove unneeded comment

* chore: fresh bun install

* test: check for type of Change Request

* test: Update agreement review test to include status change requests at the card level

- scaffolds tests for review CRs at Card Level

* chore: Update agreement review test to include status change requests at the card level

* style: :lipstick: updates content for Alerts and Modals

team collab session to help identify and update changes

* test: updates reviewAgreement test and moves create agreement test to new module

* wip: Update response status code for budget line item PATCH to 202

* test: Update review test to include status change requests at the card level

* test: Include budget change requests at the card level in review test

* test: Include budget change requests at the card level in review test

* test: Add test for reviewing budget change request date change at card level

* test: Update reviewChangeRequestsAtCardLevel.cy.js to use case-insensitive regex for status and amount checks

* fix: Update response status code in backend/openapi.yml

* fix: Update response status code in backend/openapi.yml