feat: Add toSlugCase utility function and update ReviewCard component
The code changes in this commit add a new utility function `toSlugCase` to the `utils.js` file. This function converts a string to slug case. Additionally, the `ReviewCard` component in the `ReviewCard.jsx` file is updated to import and use the `toSlugCase` function. This change allows for generating a slug case URL parameter when navigating to the `/agreements/approve/${agreementId}` route.