feat: ✨ Add scroll-to-top functionality in App component (#2543)
* feat: Add scroll-to-top functionality in App component

The App component in `App.jsx` has been updated to include scroll-to-top functionality using the `useLocation` hook from `react-router-dom`. This change ensures that when the user navigates to a new page, the window is scrolled to the top.

* chore: Remove unnecessary scroll-to-top code

The scroll-to-top functionality in the `TablePageLayout` component and the `Agreement` component has been removed. This code was redundant and no longer necessary since the `useEffect` hook was removed from both components in previous commits.

* chore: Remove unnecessary import in TablePageLayout component

The import statement for the `React` module has been removed from the `TablePageLayout.jsx` file. This import was unnecessary since the `React` module was not used in the component.