feat: ✨ Approve/decline a single agreement round 2 (#2558)
* refactor: Update TermTag component to support custom styling

The TermTag component has been updated to support custom styling by adding a new `className` prop. This will aide with layout of the Review Cards.

* refactor: offload logic to react hooks module

* test: Update agreement names in Cypress tests and Review page

The agreement names in the Cypress tests and Review page have been updated to reflect the correct values. This ensures that the tests and UI display accurate information.

* chore: Add ACTION_TYPES constant to ApproveAgreement hooks

The `ACTION_TYPES` constant has been added to the `ApproveAgreement.hooks.js` file to define the action types for approval processes. This will help improve code readability and maintainability.

* refactor: Update ReviewChangeRequestAccordion component

This commit refactors the ReviewChangeRequestAccordion component by adding a new condition to render the StatusChangeReviewCard component. The condition checks if the change request has a status change and if the change type is "STATUS". Additionally, it checks if the requested change data status matches the change request status change. This ensures that the StatusChangeReviewCard component is only rendered when the conditions are met, improving code readability and maintainability.

* docs: adds JSDocs

adds JSDocs and new constant to handle CANCEL actions

* feat: starts approveChangeRequests batch calls

adds batch api calls for budget changes.

* feat: Add change request requestor_notes and prettier

* refactor: Refactor handleDecline function and remove unused code

This commit refactors the handleDecline function in the ApproveAgreement.hooks.js file. The function has been simplified and unnecessary code has been removed. This improves code readability and maintainability.

* refactor: Update handleApproveChangeRequests function

This commit updates the handleApproveChangeRequests function in the ApproveAgreement.hooks.js file. The function now accepts an action parameter with two possible values: 'APPROVE' or 'REJECT'. This change allows for more flexibility in handling change requests. It improves code readability and maintainability.

* test: adds tests to approve change requests

* refactor: renames test module to emphasize approvals

* test: Update test module name to reflect change request approvals

- adds tests for declining change requests