fix: DRAFT BLIs before PATCHing (#2582)
* fix: DRAFT BLIs before PATCHing

* test: start new test for bug recreation

* refactor: use async/await and try catch blocks for better error handling

* feat: Add isSaving flag to useCreateBLIsAndSCs hook

* feat: Add error handling and validation to handleEditBLI function

Refactor the handleEditBLI function in CreateBLIsAndSCs.hooks.js to include error handling and validation. This ensures that the function properly handles cases where tempBudgetLines is not defined or not an array, and where budgetLineBeingEdited is an invalid index. Additionally, initialize financialSnapshot as an empty object to prevent potential errors. This improves the robustness and reliability of the code.

* docs: adds JSDocs

* refactor: Initialize financialSnapshot as an empty object in useCreateBLIsAndSCs hook

The financialSnapshot state in the useCreateBLIsAndSCs hook was not properly initialized, leading to potential errors. This commit refactors the code to initialize financialSnapshot as an empty object, ensuring its proper initialization and improving the reliability of the code.

* test: recreates scenario where bug would be present

* chore: organize imports

* fix: ensures that validation is passing

before allowing action toggle to be selected

* chore: organize imports

* fix: replaces display_name for id

* chore: Reorder imports in ServicesComponentForm/suite.js