feat: ✨ (4) 2349:  Approve Agreement toggle action  (#2585)
* refactor: remove AgreementChangesAccordion

this component is not show in the latest designs

* refactor: in-line style over css import

refine for delete-ability

* feat: adds Document Collection

renders conditionally only for Status Changes to Executing

* feat: starts BLIDiffTable and friends. Copied BudgetLinesTable as a starting point.

* refactor: Simplify BLIDiffRow component

* test: Update BLIDiffRow test to use mock functions and improve readability

* chore: organize imports

* refactor: Remove unused props and simplify BLIDiffRow component

* chore: Update AgreementTabs module CSS for notification circle styling

* feat: Add changeType and statusChangeTo props to BLIDiffTable

The BLIDiffTable component now accepts the changeType and statusChangeTo props. This allows for better customization and handling of budget line changes in different scenarios.

* chore: org imports

* style: better spacing

* fix: only show statusTo notes

- not all status type notes

* feat: filter by status of statusChangeTo

- don't show all status changes only the ones with the same statusTo

* refactor: Improve rendering of change values in BLIDiffTable

The code changes include importing the `KEY_NAMES` constant from the `ChangeRequests.constants` module in the `changeRequests.helpers.js` file.

* refactor: BLIDiffRow to improve change request filtering

The code changes in `BLIDiffRow.jsx` include refactoring the logic for filtering change requests based on the type of change (budget or status). The functions `getBudgetChangeRequests` and `getStatusChangeRequests` have been introduced to handle the filtering and retrieval of requested change data. This improves the readability and maintainability of the code.

* refactor: Improve change request filtering in BLIDiffRow

* style: Adjust size of circle badge in AgreementTabs.module.css

* refactor: Improve error handling and loading state in ApproveAgreement

The code changes in `ApproveAgreement.hooks.js` and `ApproveAgreement.jsx` include improving the error handling and loading state for the agreement. The `errorAgreement` and `isLoadingAgreement` properties have been added to handle error states and loading states respectively. This ensures a better user experience by displaying appropriate messages when there are errors or while the data is being loaded.

* docs: adds more types on return values

* feat: adds diff styles for BLIDiffTable

* chore: remove unused props

* chore: add todos for to ask UXers about

* refactor: Update statusToClassName function to use BLI_STATUS constants

The statusToClassName function in utils.js has been updated to use the BLI_STATUS constants instead of hard-coded strings for the status values. This improves code readability and maintainability by centralizing the status values in the BLI_STATUS object.

* feat: Add isApprovePage prop to AgreementCANReviewAccordion

The `AgreementCANReviewAccordion` component now accepts an `isApprovePage` prop, which is a boolean flag indicating if the page is the approve page. This prop is used to conditionally render the "After Approval" toggle button.

* feat: Add isApprovePage prop to AgreementBLIAccordion and AgreementCANReviewAccordion

This commit adds the `isApprovePage` prop to both the `AgreementBLIAccordion` and `AgreementCANReviewAccordion` components. The `isApprovePage` prop is a boolean flag that indicates if the page is the approve page. This prop is used to conditionally render the "After Approval" toggle button in both components.

* wip: toggle working in DiffTable and BLIAccordion

* chore: Add getCanById query to opsAPI

This commit adds the `getCanById` query to the `opsAPI` module in `opsAPI.js`. The query allows fetching a specific can by its ID. This query is useful for retrieving detailed information about a specific can.

* feat: Update `createUpdatedBudgetLines` function to include CAN information

The `createUpdatedBudgetLines` function in `ApproveAgreement.hooks.js` has been updated to include CAN information when applying requested changes to the budget lines. This ensures that the CAN information is properly updated when there is a requested change related to the CAN.

* feat: Include updated budget lines in AgreementBLIAccordion

The AgreementBLIAccordion component now accepts an updatedBudgetLines prop, which is an array of updated budget lines. This prop is used to display the updated budget lines when the page is in the approve mode. This change ensures that the budget lines are properly updated and displayed in the accordion component.

* refactor: Update useChangeRequests.hooks.js to improve change request handling

This commit refactors the `getFilteredChangeRequestsFromBudgetLines` function in `useChangeRequests.hooks.js` to improve the handling of change requests. It ensures that the requested changes related to budget lines are properly displayed in the `changeRequestsMessages` set. Additionally, it fixes a conditional statement to correctly handle budget changes. These changes enhance the functionality and reliability of the code.

* refactor: Improve change request handling in useChangeRequests.hooks.js

Include requested changes related to budget lines in the changeRequestsMessages set and fix conditional statement for budget changes. Enhance functionality and reliability of the code.

* revert: fixes test

* feat: Improve change request handling in useChangeRequests.hooks.js

This commit improves the handling of change requests in the `useChangeRequests.hooks.js` file. It includes requested changes related to budget lines in the `changeRequestsMessages` set and fixes a conditional statement for budget changes. These enhancements enhance the functionality and reliability of the code.

* refactor: Update useApproveAgreement.hooks.js to improve checkbox text handling

This commit updates the `useApproveAgreement.hooks.js` file to improve the handling of checkbox text based on the `changeRequestType` and `statusChangeTo` variables. It ensures that the checkbox text is correctly displayed for budget and status change requests. This enhancement improves the clarity and usability of the code.

* test: adds test data for BLI with budget change

* test: adds more robust tests

* refactor: moves logic to helpers files

* docs: add notes for future work

* chore: organize imports

- adds note for future work

* feat: Add error handling for missing budget line

The code changes in `BLIDiffRow.jsx` introduce error handling for cases where the `budgetLine` is not present. This prevents rendering errors and displays an error message instead.

* refactor: Highlight changed fields in BLIDiffRow component

This commit refactors the `BLIDiffRow` component to highlight the changed fields based on the type of change. It introduces separate highlighting for budget change and status change requests. Additionally, it adds error handling for cases where the `budgetLine` is not present, preventing rendering errors and displaying an error message instead. The code changes improve the clarity and usability of the component.

* test: Highlight changed fields in BLIDiffRow component

* refactor: Fix formatting issue in date display

The code changes fix a formatting issue in the date display for the "Date Needed" field in the `approveChangeRequestsAtAgreementLevel.cy.js` and `useChangeRequests.hooks.js` files. The double space after "Needed" has been removed to ensure consistent formatting.